Move TestNativeContextMiddleware to mux_test.go
